Removed hover trashcan and moved the deletion of individual draggables to the popout menu. Changed the color of the delete button in the popout menu aswell
