+++ TrafikkApp/components/draggable/Draggable.js
@@ -35,12 +35,10 @@ const buttonSize = 25;
  */
 const Draggable = React.memo((props) => {
     //States
-    const { source, dropZoneValues, setDraggableMoving } = props;
+    const { source } = props;
     const [imgScale, setimgScale] = useState(new Animated.Value(1));
-    const [isScaling, setIsScaling] = useState(false);
     const [tintColor, setTintColor] = useState(props.tintColor);
     const [popoutActive, setPopoutActive] = useState(false);
-    // const [popoutScaling, setPopoutScaling] = useState(new Animated.Value(1));
 
     /**
      * useEffect that is triggered when tintColor is changed
@@ -60,7 +58,6 @@ const Draggable = React.memo((props) => {
     const onDragStart = (gesture) => {
         //Start spring animation (user feedback)
         setPopoutActive(false);
-        setDraggableMoving(true);
 
         Animated.spring(imgScale, {
             toValue: 1.2,
@@ -69,20 +66,6 @@ const Draggable = React.memo((props) => {
         }).start();
     };
 
-    // /**
-    //  * While the user is dragging the object
-    //  * it will check if the object is in the view
-    //  * of the trashcan and then send feedback to trashcan
-    //  * @param {Event} gesture
-    //  */
-    // const onDragMove = (gesture) => {
-    //     if (isDropArea(gesture) && !isScaling) {
-    //         props.onTrashHover(true);
-    //     } else {
-    //         props.onTrashHover(false);
-    //     }
-    // };
-
     /**
      * When dragging event has ended, the
      * hoved animation will end and pop back to it's
@@ -90,26 +73,11 @@ const Draggable = React.memo((props) => {
      * @param {Event} gesture
      */
     const onDragEnd = (gesture) => {
-        //If in dropzone, delete element
-        if (isDropArea(gesture) && !isScaling) {
-            //End spring animation to trashcan
-            Animated.spring(imgScale, {
-                toValue: 0.1,
-                velocity: 1,
-                bounciness: 0,
-                useNativeDriver: true,
-            }).start(() => removeItem());
-        } else {
-            //End spring animation
-            Animated.spring(imgScale, {
-                toValue: 1,
-                friction: 3,
-                useNativeDriver: true,
-            }).start();
-        }
-
-        setIsScaling(false);
-        setDraggableMoving(false);
+        Animated.spring(imgScale, {
+            toValue: 1,
+            friction: 3,
+            useNativeDriver: true,
+        }).start();
     };
 
     /**
@@ -119,25 +87,7 @@ const Draggable = React.memo((props) => {
      */
     const removeItem = useCallback(() => {
         props.onRemoveItem(props.id);
-        // props.onTrashHover(false);
     });
-
-    /**
-     * Checks if the user has dropped the draggable object on top of the
-     * trashcan or not
-     * @param {Event} gesture
-     * @returns true or false if the item is inside the dropzoneArea (trashcna)
-     */
-    const isDropArea = (gesture) => {
-        var dz = dropZoneValues;
-
-        var isInZone =
-            gesture.nativeEvent.pageX >= dz.x &&
-            gesture.nativeEvent.pageX <= dz.x + dz.height &&
-            gesture.nativeEvent.pageY >= dz.y;
-
-        return isInZone;
-    };
 
     return (
         <Gestures
@@ -146,9 +96,7 @@ const Draggable = React.memo((props) => {
             rotatable={true}
             style={styles.container}
             onEnd={(event) => onDragEnd(event)}
-            // onChange={(event) => onDragMove(event)}
-            onStart={(event) => onDragStart(event)}
-            onScaleStart={() => setIsScaling(true)}>
+            onStart={(event) => onDragStart(event)}>
             <View>
                 <TouchableWithoutFeedback
                     onLongPress={() => setPopoutActive(!popoutActive)}
@@ -172,10 +120,11 @@ const Draggable = React.memo((props) => {
                             array={colors}
                             setPopoutActive={setPopoutActive}
                             popoutActive={popoutActive}
-                            exitButtonPos={colors.length - 1}
+                            removeButtonPos={colors.length - 1}
                             setTintColor={setTintColor}
                             buttonSize={buttonSize}
                             itemSize={ITEM_SIZE}
+                            removeItem={removeItem}
                         />
                     </View>
                 </TouchableWithoutFeedback>
