+++ TrafikkApp/components/draggable/Popout/Popout.js
@@ -59,8 +59,6 @@ const styles = StyleSheet.create({
         alignSelf: 'center',
         alignContent: 'center',
         justifyContent: 'center',
-        // zIndex: 999,
-        // backgroundColor: 'blue',
     },
 });
