+++ TrafikkApp/components/unusedComponents/DraggableDropzone.js
